Fixing linter errors and updating README
