+++ README.md
@@ -8,17 +8,15 @@ A turn-based battle RPG where you fight various monsters.
 For the initial scope the game will feature one class (Warrior) and one enemy (Goblin).
 The game will be rendered using the Tkinter python library.
 
-# Game Flow
-
-Initial turn decided based on which character has the highest agility stat.
-Character can chose between attacking and defending. When enemy or player HP hits zero the game ends.
+## How to run
+Run the MainApp file
 
 ## Stats
 
 HP: hit points. The character health pool. When PC (player character) HP hits zero the game should trigger a fail state, conversley when the enemy NPC (non-player character) HP hits zero the player wins
 Attack: base damage value. Subtract the defense value of oppenent and then subtract from HP to get damage
 Defense: subtracts from enemy attack to determine damage to HP
-agility: determines who goes first
+Agility: adds a dodge modifier that can prevent attacks from hitting
 
 ## Misc
 
@@ -39,13 +37,3 @@ Once MVP (minimum viable product) is achieved the following features will be con
 1. python3 -m venv .venv
 2. source .venv/bin/activate
 3. (To exit virtual environment) deactivate
-
-# App Flow
-
-1. Create character. Just name for now (default to warrior)
-2. Battle starts - List encounter flavor text in info_display
-3. Determine who moves first
-4. If enemy goes first it will attack, if player goes first they can either attack or defend
-5. Defend halves the incoming damage from the opponent attack (halved from total damage output)
-6. Keep battle going until either player or enemy HP hits zero -> display battle outcome
-7. Ask player if they want to save the results. Results include timestamp, outcome, player and enemy name
